// src/schemas/opnsense.ts
import * as z from "zod"

export const leaseSchema = z.object({
  address: z.ipv4(),
  client_id: z.string().regex(/^01:([0-9a-fA-F]{2}:){5}[0-9a-fA-F]{2}$/),
  duid: z.string().optional(),
  expire: z.int(),
  hostname: z.string(),
  hwaddr: z.string().regex(/^([0-9a-fA-F]{2}:){5}[0-9a-fA-F]{2}$/),
  iaid: z.string().optional(),
  if: z.string(),
  if_descr: z.string(),
  if_name: z.string(),
  is_reserved: z.array(z.any()),
  mac_info: z.string().optional(),
  prefix_len: z.int(),
  state: z.int(),
  type: z.string().optional(),
  valid_lifetime: z.int(),
})

export const reservationSchema = z.object({
  "%subnet": z.string().regex(/^192\.168\.80\.0\/24$/),
  client_id: z.string().regex(/^01:([0-9a-fA-F]{2}:){5}[0-9a-fA-F]{2}$/),
  description: z.string().optional(),
  hostname: z.string(),
  hw_address: z.string().regex(/^([0-9a-fA-F]{2}:){5}[0-9a-fA-F]{2}$/),
  ip_address: z.ipv4(),
  next_server: z.string().optional(),
  option: z.string().optional(),
  "option_data.boot_file_name": z.string().optional(),
  "option_data.classless_static_route": z.string().optional(),
  "option_data.domain_name": z.string().optional(),
  "option_data.domain_name_servers": z.string().optional(),
  "option_data.domain_search": z.string().optional(),
  "option_data.ntp_servers": z.string().optional(),
  "option_data.routers": z.string().optional(),
  "option_data.static_routes": z.string().optional(),
  "option_data.tftp_server_name": z.string().optional(),
  "option_data.time_servers": z.string().optional(),
  subnet: z.string(),
  uuid: z.string(),
})

export const subnetSchema = z.object({
  allocator: z.string().optional(),
  ddns_conflict_resolution_mode: z.string().optional(),
  ddns_dns_port: z.string().optional(),
  ddns_dns_server: z.string().optional(),
  ddns_domain_key_algorithm: z.string().optional(),
  ddns_domain_key_name: z.string().optional(),
  ddns_domain_key_secret: z.string().optional(),
  ddns_forward_zone: z.string().optional(),
  ddns_override_client_update: z.string(),
  ddns_override_no_update: z.string(),
  ddns_qualifying_suffix: z.string().optional(),
  ddns_reverse_zone: z.string().optional(),
  ddns_update_on_renew: z.string(),
  description: z.string().optional(),
  "match-client-id": z.string(),
  next_server: z.string().optional(),
  option: z.string().optional(),
  option_data_autocollect: z.string(),
  "option_data.boot_file_name": z.string().optional(),
  "option_data.classless_static_route": z.string().optional(),
  "option_data.domain_name": z.string().optional(),
  "option_data.domain_name_servers": z.ipv4(),
  "option_data.domain_search": z.string().optional(),
  "option_data.ntp_servers": z.ipv4(),
  "option_data.routers": z.ipv4(),
  "option_data.static_routes": z.string().optional(),
  "option_data.tftp_server_name": z.string().optional(),
  "option_data.time_servers": z.string().optional(),
  "option_data.v4_dnr": z.string().optional(),
  "option_data.v6_only_preferred": z.string().optional(),
  pools: z.string().regex(/^192\.168\.80\.50 - 192\.168\.80\.199$/),
  subnet: z.string().regex(/^192\.168\.80\.0\/24$/),
  uuid: z.string(),
  valid_lifetime: z.string().optional(),
})

export type Lease = z.infer<typeof leaseSchema>
export type Reservation = z.infer<typeof reservationSchema>
export type Subnet = z.infer<typeof subnetSchema>